# generated/api/v1/authprovider_service.swagger.json
{
  "swagger": "2.0",
  "info": {
    "title": "api/v1/authprovider_service.proto",
    "version": "version not set"
  },
  "consumes": [
    "application/json"
  ],
  "produces": [
    "application/json"
  ],
  "paths": {
    "/v1/authProviders": {
      "get": {
        "operationId": "AuthProviderService_GetAuthProviders",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/v1GetAuthProvidersResponse"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/runtimeError"
            }
          }
        },
        "parameters": [
          {
            "name": "name",
            "in": "query",
            "required": false,
            "type": "string"
          },
          {
            "name": "type",
            "in": "query",
            "required": false,
            "type": "string"
          }
        ],
        "tags": [
          "AuthProviderService"
        ]
      },
      "post": {
        "operationId": "AuthProviderService_PostAuthProvider",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/storageAuthProvider"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/runtimeError"
            }
          }
        },
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/storageAuthProvider"
            }
          }
        ],
        "tags": [
          "AuthProviderService"
        ]
      }
    },
    "/v1/authProviders/exchangeToken": {
      "post": {
        "operationId": "AuthProviderService_ExchangeToken",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/v1ExchangeTokenResponse"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/runtimeError"
            }
          }
        },
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/v1ExchangeTokenRequest"
            }
          }
        ],
        "tags": [
          "AuthProviderService"
        ]
      }
    },
    "/v1/authProviders/{id}": {
      "get": {
        "operationId": "AuthProviderService_GetAuthProvider",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/storageAuthProvider"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/runtimeError"
            }
          }
        },
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "tags": [
          "AuthProviderService"
        ]
      },
      "delete": {
        "operationId": "AuthProviderService_DeleteAuthProvider",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/v1Empty"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/runtimeError"
            }
          }
        },
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "tags": [
          "AuthProviderService"
        ]
      },
      "put": {
        "operationId": "AuthProviderService_PutAuthProvider",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/storageAuthProvider"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/runtimeError"
            }
          }
        },
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string"
          },
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/storageAuthProvider"
            }
          }
        ],
        "tags": [
          "AuthProviderService"
        ]
      },
      "patch": {
        "operationId": "AuthProviderService_UpdateAuthProvider",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/storageAuthProvider"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/runtimeError"
            }
          }
        },
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string"
          },
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/v1UpdateAuthProviderRequest"
            }
          }
        ],
        "tags": [
          "AuthProviderService"
        ]
      }
    },
    "/v1/availableAuthProviders": {
      "get": {
        "operationId": "AuthProviderService_ListAvailableProviderTypes",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/v1AvailableProviderTypesResponse"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/runtimeError"
            }
          }
        },
        "tags": [
          "AuthProviderService"
        ]
      }
    },
    "/v1/login/authproviders": {
      "get": {
        "operationId": "AuthProviderService_GetLoginAuthProviders",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/v1GetLoginAuthProvidersResponse"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/runtimeError"
            }
          }
        },
        "tags": [
          "AuthProviderService"
        ]
      }
    }
  },
  "definitions": {
    "AuthProviderRequiredAttribute": {
      "type": "object",
      "properties": {
        "attributeKey": {
          "type": "string"
        },
        "attributeValue": {
          "type": "string"
        }
      },
      "description": "RequiredAttribute allows to specify a set of attributes which ALL are required to be returned\nby the auth provider.\nIf any attribute is missing within the external claims of the token issued by Central, the\nauthentication request to this IdP is considered failed."
    },
    "AvailableProviderTypesResponseAuthProviderType": {
      "type": "object",
      "properties": {
        "type": {
          "type": "string"
        },
        "suggestedAttributes": {
          "type": "array",
          "items": {
            "type": "string"
          }
        }
      }
    },
    "GetLoginAuthProvidersResponseLoginAuthProvider": {
      "type": "object",
      "properties": {
        "id": {
          "type": "string"
        },
        "name": {
          "type": "string"
        },
        "type": {
          "type": "string"
        },
        "loginUrl": {
          "type": "string"
        }
      }
    },
    "UserInfoResourceToAccess": {
      "type": "object",
      "properties": {
        "resourceToAccess": {
          "type": "object",
          "additionalProperties": {
            "$ref": "#/definitions/storageAccess"
          }
        }
      },
      "description": "ResourceToAccess represents a collection of permissions. It is wire\ncompatible with the old format of storage.Role and replaces it in\nplaces where only aggregated permissions are required."
    },
    "protobufAny": {
      "type": "object",
      "properties": {
        "typeUrl": {
          "type": "string",
          "description": "A URL/resource name that uniquely identifies the type of the serialized\nprotocol buffer message. This string must contain at least\none \"/\" character. The last segment of the URL's path must represent\nthe fully qualified name of the type (as in\n`path/google.protobuf.Duration`). The name should be in a canonical form\n(e.g., leading \".\" is not accepted).\n\nIn practice, teams usually precompile into the binary all types that they\nexpect it to use in the context of Any. However, for URLs which use the\nscheme `http`, `https`, or no scheme, one can optionally set up a type\nserver that maps type URLs to message definitions as follows:\n\n* If no scheme is provided, `https` is assumed.\n* An HTTP GET on the URL must yield a [google.protobuf.Type][]\n  value in binary format, or produce an error.\n* Applications are allowed to cache lookup results based on the\n  URL, or have them precompiled into a binary to avoid any\n  lookup. Therefore, binary compatibility needs to be preserved\n  on changes to types. (Use versioned type names to manage\n  breaking changes.)\n\nNote: this functionality is not currently available in the official\nprotobuf release, and it is not used for type URLs beginning with\ntype.googleapis.com.\n\nSchemes other than `http`, `https` (or the empty scheme) might be\nused with implementation specific semantics."
        },
        "value": {
          "type": "string",
          "format": "byte",
          "description": "Must be a valid serialized protocol buffer of the above specified type."
        }
      },
      "description": "`Any` contains an arbitrary serialized protocol buffer message along with a\nURL that describes the type of the serialized message.\n\nProtobuf library provides support to pack/unpack Any values in the form\nof utility functions or additional generated methods of the Any type.\n\nExample 1: Pack and unpack a message in C++.\n\n    Foo foo = ...;\n    Any any;\n    any.PackFrom(foo);\n    ...\n    if (any.UnpackTo(\u0026foo)) {\n      ...\n    }\n\nExample 2: Pack and unpack a message in Java.\n\n    Foo foo = ...;\n    Any any = Any.pack(foo);\n    ...\n    if (any.is(Foo.class)) {\n      foo = any.unpack(Foo.class);\n    }\n\nExample 3: Pack and unpack a message in Python.\n\n    foo = Foo(...)\n    any = Any()\n    any.Pack(foo)\n    ...\n    if any.Is(Foo.DESCRIPTOR):\n      any.Unpack(foo)\n      ...\n\nExample 4: Pack and unpack a message in Go\n\n     foo := \u0026pb.Foo{...}\n     any, err := anypb.New(foo)\n     if err != nil {\n       ...\n     }\n     ...\n     foo := \u0026pb.Foo{}\n     if err := any.UnmarshalTo(foo); err != nil {\n       ...\n     }\n\nThe pack methods provided by protobuf library will by default use\n'type.googleapis.com/full.type.name' as the type URL and the unpack\nmethods only use the fully qualified type name after the last '/'\nin the type URL, for example \"foo.bar.com/x/y.z\" will yield type\nname \"y.z\".\n\n\nJSON\n\nThe JSON representation of an `Any` value uses the regular\nrepresentation of the deserialized, embedded message, with an\nadditional field `@type` which contains the type URL. Example:\n\n    package google.profile;\n    message Person {\n      string first_name = 1;\n      string last_name = 2;\n    }\n\n    {\n      \"@type\": \"type.googleapis.com/google.profile.Person\",\n      \"firstName\": \u003cstring\u003e,\n      \"lastName\": \u003cstring\u003e\n    }\n\nIf the embedded message type is well-known and has a custom JSON\nrepresentation, that representation will be embedded adding a field\n`value` which holds the custom JSON in addition to the `@type`\nfield. Example (for message [google.protobuf.Duration][]):\n\n    {\n      \"@type\": \"type.googleapis.com/google.protobuf.Duration\",\n      \"value\": \"1.212s\"\n    }"
    },
    "runtimeError": {
      "type": "object",
      "properties": {
        "error": {
          "type": "string"
        },
        "code": {
          "type": "integer",
          "format": "int32"
        },
        "message": {
          "type": "string"
        },
        "details": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/protobufAny"
          }
        }
      }
    },
    "storageAccess": {
      "type": "string",
      "enum": [
        "NO_ACCESS",
        "READ_ACCESS",
        "READ_WRITE_ACCESS"
      ],
      "default": "NO_ACCESS"
    },
    "storageAuthProvider": {
      "type": "object",
      "properties": {
        "id": {
          "type": "string"
        },
        "name": {
          "type": "string"
        },
        "type": {
          "type": "string"
        },
        "uiEndpoint": {
          "type": "string"
        },
        "enabled": {
          "type": "boolean"
        },
        "config": {
          "type": "object",
          "additionalProperties": {
            "type": "string"
          }
        },
        "loginUrl": {
          "type": "string",
          "description": "The login URL will be provided by the backend, and may not be specified in a request."
        },
        "validated": {
          "type": "boolean"
        },
        "extraUiEndpoints": {
          "type": "array",
          "items": {
            "type": "string"
          },
          "description": "UI endpoints which to allow in addition to `ui_endpoint`. I.e., if a login request\nis coming from any of these, the auth request will use these for the callback URL,\nnot ui_endpoint."
        },
        "active": {
          "type": "boolean"
        },
        "requiredAttributes": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/AuthProviderRequiredAttribute"
          },
          "description": "EXPERIMENTAL."
        }
      },
      "title": "Next Tag: 9"
    },
    "storageAuthStatus": {
      "type": "object",
      "properties": {
        "userId": {
          "type": "string"
        },
        "serviceId": {
          "$ref": "#/definitions/storageServiceIdentity"
        },
        "expires": {
          "type": "string",
          "format": "date-time"
        },
        "refreshUrl": {
          "type": "string"
        },
        "authProvider": {
          "$ref": "#/definitions/storageAuthProvider"
        },
        "userInfo": {
          "$ref": "#/definitions/storageUserInfo"
        },
        "userAttributes": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/storageAuthStatusUserAttribute"
          }
        }
      }
    },
    "storageAuthStatusUserAttribute": {
      "type": "object",
      "properties": {
        "key": {
          "type": "string"
        },
        "values": {
          "type": "array",
          "items": {
            "type": "string"
          }
        }
      }
    },
    "storageServiceIdentity": {
      "type": "object",
      "properties": {
        "serialStr": {
          "type": "string"
        },
        "serial": {
          "type": "string",
          "format": "int64"
        },
        "id": {
          "type": "string"
        },
        "type": {
          "$ref": "#/definitions/storageServiceType"
        },
        "initBundleId": {
          "type": "string"
        }
      }
    },
    "storageServiceType": {
      "type": "string",
      "enum": [
        "UNKNOWN_SERVICE",
        "SENSOR_SERVICE",
        "CENTRAL_SERVICE",
        "REMOTE_SERVICE",
        "COLLECTOR_SERVICE",
        "MONITORING_UI_SERVICE",
        "MONITORING_DB_SERVICE",
        "MONITORING_CLIENT_SERVICE",
        "BENCHMARK_SERVICE",
        "SCANNER_SERVICE",
        "SCANNER_DB_SERVICE",
        "ADMISSION_CONTROL_SERVICE"
      ],
      "default": "UNKNOWN_SERVICE"
    },
    "storageUserInfo": {
      "type": "object",
      "properties": {
        "username": {
          "type": "string"
        },
        "friendlyName": {
          "type": "string"
        },
        "permissions": {
          "$ref": "#/definitions/UserInfoResourceToAccess"
        },
        "roles": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/storageUserInfoRole"
          }
        }
      }
    },
    "storageUserInfoRole": {
      "type": "object",
      "properties": {
        "name": {
          "type": "string"
        },
        "resourceToAccess": {
          "type": "object",
          "additionalProperties": {
            "$ref": "#/definitions/storageAccess"
          }
        }
      },
      "description": "Role is wire compatible with the old format of storage.Role and\nhence only includes role name and associated permissions."
    },
    "v1AvailableProviderTypesResponse": {
      "type": "object",
      "properties": {
        "authProviderTypes": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/AvailableProviderTypesResponseAuthProviderType"
          }
        }
      }
    },
    "v1Empty": {
      "type": "object"
    },
    "v1ExchangeTokenRequest": {
      "type": "object",
      "properties": {
        "externalToken": {
          "type": "string",
          "description": "The external authentication token. The server will mask the value of this credential in responses and logs."
        },
        "type": {
          "type": "string"
        },
        "state": {
          "type": "string"
        }
      }
    },
    "v1ExchangeTokenResponse": {
      "type": "object",
      "properties": {
        "token": {
          "type": "string"
        },
        "clientState": {
          "type": "string"
        },
        "test": {
          "type": "boolean"
        },
        "user": {
          "$ref": "#/definitions/storageAuthStatus"
        }
      }
    },
    "v1GetAuthProvidersResponse": {
      "type": "object",
      "properties": {
        "authProviders": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/storageAuthProvider"
          }
        }
      }
    },
    "v1GetLoginAuthProvidersResponse": {
      "type": "object",
      "properties": {
        "authProviders": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/GetLoginAuthProvidersResponseLoginAuthProvider"
          }
        }
      }
    },
    "v1UpdateAuthProviderRequest": {
      "type": "object",
      "properties": {
        "id": {
          "type": "string"
        },
        "name": {
          "type": "string"
        },
        "enabled": {
          "type": "boolean"
        }
      }
    }
  }
}